fix : 홈 수정 #120
