+++ app/components/home/StockCategory.Styled.ts
@@ -1,24 +1,28 @@
+import Link from "next/link";
 import styled from "styled-components";
+
+export const Container = styled.div`
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    max-width: 24.5rem;
+`;
 
 export const TitleBox = styled.div`
     display: inline-block;
-    padding: 0.25rem 0.875rem; /* 4px 14px */
+    width: fit-content;
+    padding: 0.25rem 0.875rem;
     background-color: var(--gray-200);
-    border-radius: 0.375rem; /* 6px */
-    font-size: 1rem; /* 유지 */
+    border-radius: 5px;
+    font-size: 1rem;
     font-weight: bold;
-`;
-
-export const Container = styled.div`
-    width: 100%;
-    max-width: 24.5rem; /* 392px */
 `;
 
 export const Grid = styled.div`
     display: grid;
     grid-template-columns: repeat(3, 1fr);
-    gap: 1rem; /* 16px */
-    margin-top: 1rem; /* 16px */
+    gap: 1rem;
+    margin-top: 1rem;
 `;
 
 export const CategoryWrapper = styled.div`
@@ -29,26 +33,25 @@ export const CategoryWrapper = styled.div`
 `;
 
 export const CategoryName = styled.span`
-    font-size: 0.875rem; /* 14px */
+    font-size: var(--font-size-base);
     font-weight: bold;
-    margin-bottom: 0.375rem; /* 6px */
+    margin-bottom: 5px; /* 위 간격 조정 */
+    cursor: default;
 `;
 
-export const CategoryCard = styled.div<{ color: string }>`
-    width: 6.5rem; /* 104px */
-    height: 6.5rem; /* 104px */
+export const CategoryCard = styled(Link)<{ color: string }>`
+    width: 6.5rem;
+    height: 6.5rem;
     background-color: ${({ color }) => color}; /* 각 카드의 배경색 */
     display: flex;
     flex-direction: column;
     align-items: center;
     justify-content: center;
-    border-radius: 1rem; /* 16px */
-    box-shadow: 0.125rem 0.125rem 0.625rem rgba(0, 0, 0, 0.1); /* 2px 2px 10px */
-    cursor: pointer;
+    border-radius: 1rem;
+    box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.1);
     transition: filter 0.2s;
 
     &:hover {
         filter: brightness(90%);
     }
 `;
-export const ImageWrapper = styled.div``;
