+++ app/components/home/StockList.tsx
@@ -1,85 +1,90 @@
-"use client";
-import {
-    StockContainer,
-    TitleBox,
-    StockItems,
-    StockLink,
-    StockItem,
-    StockInfo,
-    StockLeft,
-    LogoWrapper,
-    StockName,
-    StockRight,
-    StockPrice,
-    StockChange,
-    StockImage,
-} from "@/app/components/home/StockList.Styled";
-const stockData = [
-    {
-        id: 1,
-        name: "삼성전자",
-        logo: "/stock-logo/samsung.png",
-        price: "56,400",
-        change: "+3215",
-        changePercent: "+1.5%",
-        symbol: "삼성전자", // 주식 심볼 추가
-    },
-    {
-        id: 2,
-        name: "SK하이닉스",
-        logo: "/stock-logo/sk.png",
-        price: "198,700",
-        change: "-3215",
-        changePercent: "-1.5%",
-        symbol: "SK하이닉스",
-    },
-    {
-        id: 3,
-        name: "LG에너지솔루션",
-        logo: "/stock-logo/lg.png",
-        price: "369,500",
-        change: "+3215",
-        changePercent: "+1.5%",
-        symbol: "LG에너지솔루션",
-    },
-    {
-        id: 4,
-        name: "카카오",
-        logo: "/stock-logo/kakao.png",
-        price: "44,300",
-        change: "-3215",
-        changePercent: "-1.5%",
-        symbol: "카카오",
-    },
-];
-const StockList = () => {
-    return (
-        <StockContainer>
-            <TitleBox>인기 종목</TitleBox>
-            <StockItems>
-                {stockData.map((stock) => (
-                    <StockLink href={`/stock/${stock.symbol}`} key={stock.id}>
-                        <StockItem>
-                            <StockInfo>
-                                <StockLeft>
-                                    <LogoWrapper>
-                                        <StockImage src={stock.logo} alt={stock.name} width={40} height={40} />
-                                    </LogoWrapper>
-                                    <StockName>{stock.name}</StockName>
-                                </StockLeft>
-                                <StockRight>
-                                    <StockPrice>{stock.price} 원</StockPrice>
-                                    <StockChange $isPositive={stock.change.startsWith("+")}>
-                                        {stock.change}원 ({stock.changePercent})
-                                    </StockChange>
-                                </StockRight>
-                            </StockInfo>
-                        </StockItem>
-                    </StockLink>
-                ))}
-            </StockItems>
-        </StockContainer>
-    );
+import { LogoWrapper, StockChange, StockImage, StockItemBox, StockLeft, StockLink, StockName, StockPrice, StockRight } from "@/app/components/home/StockList.Styled";
+import { StockListResponseDto } from "@/application/usecases/home/dtos/StockListResponseDto";
+
+import { env } from "@/config/env";
+
+interface StockListProps {
+  path: string;
+}
+
+async function fetchStocks(path: string): Promise<StockListResponseDto[]> {
+  const res = await fetch(`${env.NEXT_PUBLIC_BASE_URL}${path}`, {
+    cache: "no-store",
+  });
+
+  return await res.json();
+}
+
+const StockList = async ({ path }: StockListProps) => {
+  const stockData = await fetchStocks(path);
+
+  return (
+    <>
+      <StockItemBox>
+        {stockData?.map(
+          ({
+            index,
+            stockCode,
+            stockImage,
+            stockId,
+            stockName,
+            currentPrice,
+          }) => {
+            const isPositive = !currentPrice.prdyVrss.startsWith("-");
+
+            const formattedVrss =
+              currentPrice.prdyVrss === "0"
+                ? "0"
+                : isPositive
+                ? `+${Number(currentPrice.prdyVrss).toLocaleString()}`
+                : Number(currentPrice.prdyVrss).toLocaleString();
+
+            const formattedCtrt =
+              currentPrice.prdyCtrt === "0"
+                ? "0%"
+                : isPositive
+                ? `+${currentPrice.prdyCtrt}%`
+                : `${currentPrice.prdyCtrt}%`;
+
+            return (
+              <StockLink href={`/stock/${stockCode}`} key={stockId}>
+                <StockLeft>
+                  {index && <span>{index}.</span>}
+                  <LogoWrapper>
+                    {stockImage ? (
+                      <StockImage
+                        src={`/stock/${stockImage}.png`}
+                        alt={stockName}
+                        width={40}
+                        height={40}
+                      />
+                    ): <>
+                    <StockImage
+                        src={`/stock/DEFAULT.png`}
+                        alt={"stockDefault"}
+                        width={40}
+                        height={40}
+                        />
+                    </> }
+                  </LogoWrapper>
+                  <StockName>{stockName}</StockName>
+                </StockLeft>
+
+                <StockRight>
+                  <StockPrice>
+                    {Number(currentPrice.stckPrpr).toLocaleString()} 원
+                  </StockPrice>
+                  <StockChange $isPositive={isPositive}>
+                    {formattedVrss}원 ({formattedCtrt})
+                  </StockChange>
+                </StockRight>
+              </StockLink>
+            );
+          }
+        )}
+      </StockItemBox>
+    </>
+  );
 };
 
 export default StockList;
