--- app/te/page.tsx
@@ -1,11 +1,0 @@
-import Te from "@/app/te/components/Te";
-
-const TePage = () => {
-  const path = `/api/home`;
-  return (
-    <>
-      <Te path={path}/>
-    </>
-  );
-}
-export default TePage;
